% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawallometry.R
\name{drawallometry}
\alias{drawallometry}
\title{Draw allometric equation}
\usage{
drawallometry(species, type = "total")
}
\arguments{
\item{species}{the specie or functional group to be drawn}

\item{type}{either 'total'  (total fuel) of 'fine' (fine fuels)}
}
\description{
Draw allometric equation
}
\examples{

}
