% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual2species.R
\name{individual2species}
\alias{individual2species}
\title{Transform individual data to species data}
\usage{
individual2species(x, sampledarea = 20, maxcover = 100)
}
\arguments{
\item{x}{data frame with columns 'plot', 'species', 'H' (height in cm), 'D1' and 'D2' (in cm)}

\item{sampledarea}{sampled area in squared meters}

\item{maxcover}{maximum allowed cover (set to NA to avoid truncation)}
}
\value{
data frame with columns 'plot', 'species', 'H' (mean height in cm), 'C' (cover in percent)
}
\description{
Transform individual data to species data
}
\examples{
plot = c(1,1,2,2,2)
species = c("Erica arborea","Cistus albidus", "Erica arborea", "Chamaerops humilis", "Unknown")
H = c(60,200,100,250,100)
D1 = c(10,100,30, 50,25)
D2 = D1
x = data.frame(plot, species, H, D1, D2)

individual2species(x, sampledarea = 2)
}
