% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantfuelbiomass.R
\name{plantfuelbiomass}
\alias{plantfuelbiomass}
\title{Plant fuel biomass}
\usage{
plantfuelbiomass(x, type = "total", agg = "none", customParams = NULL,
  na.rm = TRUE)
}
\arguments{
\item{x}{data frame with columns 'plot', 'individual', 'species', 'H' (height in cm), 'D1' and 'D2' (in cm)}

\item{type}{either 'total'  (total fuel) or 'fine' (fine fuels)}

\item{agg}{aggregation of results. Either 'none', 'species' or 'plot'}

\item{customParams}{custom allometry parameter table (for species not in default params)}

\item{na.rm}{whether to exclude missing values when aggregating biomass}
}
\value{
a vector of dry weight (biomass) in kg
}
\description{
Calculates dry weight (biomass, in kg) of total or fine fuels corresponding to plant data
}
\examples{
plot = c(1,1,2,2,2)
individual = c(1,2,1,2,3)
species = c("Erica arborea","Cistus albidus", "Erica arborea", "Chamaerops humilis", "Unknown")
H = c(60,200,100,250,100)
D1 = c(10,100,30, 50,25)
D2 = D1
x = data.frame(plot, individual, species, H, D1, D2)

plantfuelbiomass(x)
}
