% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrubfuelbiomass.R
\name{shrubfuelbiomass}
\alias{shrubfuelbiomass}
\title{Shrub fuel biomass}
\usage{
shrubfuelbiomass(x, type = "total", allometric = TRUE, excludeSSP = TRUE,
  agg = "none", customParams = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{data frame with columns 'plot', 'species', 'H' (height in cm), 'D1' and 'D2' (in cm)}

\item{type}{either 'total'  (total fuel) of 'fine' (fine fuels)}

\item{allometric}{wether to use allometric equations or bulk density estimates}

\item{excludeSSP}{excludes subspecies information for species matching}

\item{agg}{aggregation of results. Either 'none', 'species', 'speciesplot', 'plotspecies' or 'plot'}

\item{customParams}{custom allometry parameter table (for species not in default params)}

\item{na.rm}{whether to exclude missing values when aggregating biomass}
}
\value{
a vector of dry weight (biomass) in kg
}
\description{
Calculates dry weight (biomass, in kg) of total or fine fuels corresponding to individual shrub data
}
\examples{
plot = c(1,1,2,2,2)
species = c("Erica arborea","Cistus albidus", "Erica arborea", "Chamaerops humilis", "Unknown")
H = c(60,200,100,250,100)
D1 = c(10,100,30, 50,25)
D2 = D1
x = data.frame(plot, species, H, D1, D2)

shrubfuelbiomass(x)
}
