% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrubindividual2species.R
\name{shrubindividual2species}
\alias{shrubindividual2species}
\title{Transform shrub individual data to species data}
\usage{
shrubindividual2species(x, sampledarea = 20, maxcover = 100, na.rm = TRUE)
}
\arguments{
\item{x}{data frame with columns 'plot', 'species', 'H' (height in cm), 'D1' and 'D2' (in cm)}

\item{sampledarea}{sampled area in squared meters}

\item{maxcover}{maximum allowed cover (set to NA to avoid truncation)}

\item{na.rm}{whether to exclude missing values when averaging heights and adding areas}
}
\value{
data frame with columns 'plot', 'species', 'H' (mean height in cm), 'C' (cover in percent)
}
\description{
Transforms shrub individual data (plant heights and diameters) into species level data (average height and cover).
}
\examples{
plot = c(1,1,2,2,2)
species = c("Erica arborea","Cistus albidus", "Erica arborea", "Chamaerops humilis", "Unknown")
H = c(60,200,100,250,100)
D1 = c(10,100,30, 50,25)
D2 = D1
x = data.frame(plot, species, H, D1, D2)

shrubindividual2species(x, sampledarea = 2)
}
