% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrubspeciesfuel.R
\name{shrubspeciesfuel}
\alias{shrubspeciesfuel}
\title{Shrub species fuel loading}
\usage{
shrubspeciesfuel(x, type = "total", allometric = TRUE, excludeSSP = TRUE,
  var = FALSE, agg = "none", customParams = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{data frame with columns 'plot', 'species', 'H' (mean height in cm), 'C' (cover in percent)}

\item{type}{either 'total'  (total fuel) or 'fine' (fine fuels)}

\item{allometric}{wether to use allometric equations or bulk density estimates}

\item{excludeSSP}{excludes subspecies information for species matching}

\item{var}{a flag to indicate that variance of estimates is desired}

\item{agg}{aggregation of results. Either 'none' or 'plot'}

\item{customParams}{custom allometry parameter table (for species not in default params)}

\item{na.rm}{whether to exclude missing values when aggregating loading}
}
\value{
a vector with loading values (kg/m2)
}
\description{
Calculates loading (kg/m2) of total or fine fuels corresponding to shrub species data
}
\examples{
plot = c(1,1,2,2,2)
species = c("Erica arborea","Cistus albidus", "Erica arborea", "Chamaerops humilis", "Unknown")
H = c(60,200,100,250,100)
C = c(10,100,30, 50,25)
x = data.frame(plot, species, H, C)

shrubspeciesfuel(x)
}
