% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationCombuscat.R
\docType{data}
\name{combuscat}
\alias{combuscat}
\alias{individual_data,}
\alias{group_params_area}
\alias{group_params_fine}
\alias{group_params_total}
\alias{sp_params_area}
\alias{sp_params_fine}
\alias{sp_params_total}
\title{Allometric and bulk density parameter set from project COMBUSCAT}
\description{
Description of the aims of the project and funding.
}
\details{
The data set includes seven data items:

\itemize{
 \item individual_data Individual-based data underlying the development of allometric equations
 \item sp_params_area Species-specific parameter for the relation between height (in cm) and area (in cm2)
 \item sp_params_fine Species-specific allometric parameters for the relation between volume (in m3) and dry weight (in kg) of fine fuels
 \item sp_params_total Species-specific allometric parameters for the relation between volume (in m3) and dry weight (in kg) of total fuels
 \item group_params_area Group-specific parameter for the relation between height (in cm) and area (in cm2)
 \item group_params_fine Group-specific allometric parameters for the relation between volume (in m3) and dry weight (in kg) of fine fuels
 \item group_params_total Group-specific allometric parameters for the relation between volume (in m3) and dry weight (in kg) of total fuels
}
}
\references{
Project reference.
}
\author{
Parameter set derived from project COMBUSCAT.
}
\keyword{data}
