% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawallometry.R
\name{drawallometry}
\alias{drawallometry}
\title{Draw allometric equation}
\usage{
drawallometry(species, type = "total", log = "")
}
\arguments{
\item{species}{the specie or functional group to be drawn}

\item{type}{'total'  (total fuel), 'fine' (fine fuels), or 'both'}
}
\description{
Draw allometric equation
}
\examples{
drawallometry("Cistus albidus")
}
