% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawshruballometry.R
\name{drawshruballometry}
\alias{drawshruballometry}
\title{Draw shrub allometric equation}
\usage{
drawshruballometry(species, type = "total", log = "")
}
\arguments{
\item{species}{The shrub species or functional group to be drawn}

\item{type}{Fuel type to draw: 'total'  (total fuel), 'fine' (fine fuels), or 'both'}

\item{log}{Specifies logarithmic axis: 'x', 'y' or 'xy'}
}
\description{
Draw shrub allometric equation
}
\examples{
drawshruballometry("Cistus albidus")
}
