% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individualshrub2species.R
\name{individualshrub2species}
\alias{individualshrub2species}
\title{Transform shrub individual data to species data}
\usage{
individualshrub2species(x, sampledarea = 20, maxcover = 100, na.rm = TRUE,
  sd.H = FALSE)
}
\arguments{
\item{x}{data frame with columns 'plot', 'species', 'H' (height in cm), 'D1' and 'D2' (in cm)}

\item{sampledarea}{sampled area in squared meters}

\item{maxcover}{maximum allowed cover (set to NA to avoid truncation)}

\item{na.rm}{whether to exclude missing values when averaging heights and adding areas}

\item{sd.H}{whether to include the standard deviation of heights in the output}
}
\value{
data frame with columns 'plot', 'species', 'H' (mean height in cm), 'C' (cover in percent)
}
\description{
Transforms shrub individual data (plant heights and diameters) into species level data (average height and cover).
}
\examples{
plot = c(1,1,1,1,1)
species = c("Erica arborea","Cistus albidus", "Erica arborea", "Cistus albidus", "Cistus albidus")
H = c(200,50,100,40,30)
D1 = c(140,40,100, 35,30)
D2 = D1
x = data.frame(plot, species, H, D1, D2)

individualshrub2species(x, sampledarea = 4)
}
